error checking for popup when adding to collection from wishlist

+++ MD3/app/src/main/java/edu/wit/mobilepp/md3/CardItemMakeupWishListAdapter.java
@@ -6,6 +6,7 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.TextInputEditText;
+import android.support.design.widget.TextInputLayout;
 import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -130,6 +131,8 @@ public class CardItemMakeupWishListAdapter extends ArrayAdapter<CardItemMakeupWi
 
                     TextInputEditText purchase_date = (TextInputEditText)customView.findViewById(R.id.purchase_date2);
                     TextInputEditText lifespan = (TextInputEditText)customView.findViewById(R.id.lifespan2);
+
+                    TextInputLayout date_layout = (TextInputLayout)customView.findViewById(R.id.textInputLayout6);
                     @Override
                     public void onClick(View v) {
                         DateFormat dateFormat = new SimpleDateFormat("MM/dd/yyyy");
@@ -139,7 +142,7 @@ public class CardItemMakeupWishListAdapter extends ArrayAdapter<CardItemMakeupWi
                             date = dateFormat.format(date1);
                             Log.v("new",date1.toString());
                         }catch (ParseException e){
-                            date="a";
+                            date = "";
                             e.printStackTrace();
                         }
                         Integer life;
@@ -238,33 +241,41 @@ public class CardItemMakeupWishListAdapter extends ArrayAdapter<CardItemMakeupWi
 //                        }
 
                         Log.v("wl", date);
-                        String path = "/data/data/" + getContext().getPackageName() + "/makeup_wishlist.db";
-                        SQLiteDatabase db;
-                        db = SQLiteDatabase.openOrCreateDatabase(path, null);
-                        String path2 = "/data/data/" + getContext().getPackageName() + "/makeup_collection.db";
-                        SQLiteDatabase db2;
-                        db2 = SQLiteDatabase.openOrCreateDatabase(path2, null);
-                        // Create a table - people
-                        String sql = "CREATE TABLE IF NOT EXISTS makeup_collection" +
-                                "(_id INTEGER PRIMARY KEY AUTOINCREMENT, brand TEXT, product TEXT, category TEXT, shade TEXT, date TEXT, life INTEGER, days INTEGER, date_sort TEXT);";
-                        db2.execSQL(sql);
-                        ContentValues values = new ContentValues();
-                        values.put("brand", item.brand);
-                        values.put("product", item.product);
-                        values.put("category", item.category);
-                        values.put("shade", item.shade);
-                        values.put("date", date);
-                        values.put("life", life);
-                        //values.put("days", days_left);
-                        db2.insert("makeup_collection", null, values);
-                        db2.execSQL("UPDATE makeup_collection SET date_sort=substr(date,7,4)||'-'||substr(date,1,2)||'-'||substr(date,4,2);");
-                        db.execSQL("DELETE FROM makeup_wishlist WHERE _ID= " + item.id);
-                        l.remove(position);
-                        notifyDataSetChanged();
-                        //Close the database
-                        db2.close();
-                        db.close();
-                        popupWindow.dismiss();
+
+                        //error check
+                        date_layout.setError(null);
+
+                        if(date.isEmpty()) {
+                            date_layout.setError("You must enter a purchase date.");
+                        } else {
+                            String path = "/data/data/" + getContext().getPackageName() + "/makeup_wishlist.db";
+                            SQLiteDatabase db;
+                            db = SQLiteDatabase.openOrCreateDatabase(path, null);
+                            String path2 = "/data/data/" + getContext().getPackageName() + "/makeup_collection.db";
+                            SQLiteDatabase db2;
+                            db2 = SQLiteDatabase.openOrCreateDatabase(path2, null);
+                            // Create a table - people
+                            String sql = "CREATE TABLE IF NOT EXISTS makeup_collection" +
+                                    "(_id INTEGER PRIMARY KEY AUTOINCREMENT, brand TEXT, product TEXT, category TEXT, shade TEXT, date TEXT, life INTEGER, days INTEGER, date_sort TEXT);";
+                            db2.execSQL(sql);
+                            ContentValues values = new ContentValues();
+                            values.put("brand", item.brand);
+                            values.put("product", item.product);
+                            values.put("category", item.category);
+                            values.put("shade", item.shade);
+                            values.put("date", date);
+                            values.put("life", life);
+                            //values.put("days", days_left);
+                            db2.insert("makeup_collection", null, values);
+                            db2.execSQL("UPDATE makeup_collection SET date_sort=substr(date,7,4)||'-'||substr(date,1,2)||'-'||substr(date,4,2);");
+                            db.execSQL("DELETE FROM makeup_wishlist WHERE _ID= " + item.id);
+                            l.remove(position);
+                            notifyDataSetChanged();
+                            //Close the database
+                            db2.close();
+                            db.close();
+                            popupWindow.dismiss();
+                        }
                     }
                 });
